Monitoring/app/requirements.txt
streamlit
psycopg2-binary
lancedb
pandas
numpy
pypdf
pyarrow
sentence-transformers
transformers
langchain
langchain-community
